requirements.txt
pexels-api
requests~=2.32.5
python-dotenv~=1.2.1
Flask~=3.1.2